requirements.txt
pandas
torch
transformers==4.55.2
scikit-learn
datasets
evaluate
accelerate>=0.26.0
flask
flask-login
flask-sqlalchemy
flask-wtf
rapidfuzz
Pillow
